Update bug in http-error & implemented controller - Updated http-error.ts to utilize getters & setters - Created places-controller.ts to handle middleware functions - Refactor places-routes to utilize places-controller.ts - Updated app.ts to use appropiate getter for http-error.ts
